test(resident): os0 fast cases assert the resident engines ran; Pmax compared in float32
Under the resident flag set the K=1 replay, local replay, cold starts and
perturbation replay now fail if their global or local pass fell back to the
compact or exact local engine: their RELION gates alone cannot tell which
engine produced the maps. The os0 unit test compares the carried coarse Pmax in
the scoring precision.

+++ tests/unit/test_resident_zero_oversampling.py
@@ -199,9 +199,10 @@ def test_zero_oversampling_resident_driver_matches_the_compact_engine(_resident_
     # The coarse winner and Pmax are carried through, not recomputed.
     np.testing.assert_array_equal(compact.hard_assignment, resident.hard_assignment)
     np.testing.assert_array_equal(compact.best_rotation_indices, resident.best_rotation_indices)
+    # Published in the scoring precision (float32), as RELION's max_weight / sum_weight.
     assert_matches(
-        np.asarray(resident.relion_stats.max_posterior_per_image, dtype=np.float64),
-        np.asarray(args["relion_coarse_max_posterior"], dtype=np.float64),
+        np.asarray(resident.relion_stats.max_posterior_per_image, dtype=np.float32),
+        np.asarray(args["relion_coarse_max_posterior"], dtype=np.float32),
     )
     for field in (
         "log_evidence_per_image",
